Formatting fixes, more warnings removed, install script fixes

+++ pswritermake.sh
@@ -3,4 +3,11 @@
 Warning_FLAGS="-Wall -Wextra -Wpedantic -Wformat=2 -Wno-unused-parameter -Wshadow -Wwrite-strings -Wstrict-prototypes -Wold-style-definition -Wredundant-decls -Wnested-externs -Wmissing-include-dirs -Wjump-misses-init -Wlogical-op"
 
 gcc -O3 ${Warning_FLAGS} pswriter.c -lm
-cp a.out pswriter
+
+if [ -e a.out ]
+then
+	cp a.out pswriter
+	rm a.out
+else
+	echo "Something went wrong, 'a.out' not produced"
+fi
